bibliotheque_project/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.1.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==22.0.0